# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API da Livraria Libris",
    "description": "Documentação da API da livraria online Libris",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/clientes/": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Consulta um cliente",
        "description": "Consulta os dados de um cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "email": {
                  "type": "string",
                  "example": "joao@email.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Cliente consultado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Cadastra um cliente",
        "description": "Realiza o cadastro de um novo cliente.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "email": {
                  "type": "string",
                  "example": "joao@email.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Cliente cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
    "/clientes/{id}": {
      "delete": {
        "tags": [
          "Clientes"
        ],
        "summary": "Exclui um cliente",
        "description": "Exclui um cliente existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "email": {
                  "type": "string",
                  "example": "joao@email.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente excluído com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualiza um cliente",
        "description": "Atualiza os dados de um cliente existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "João da Silva"
                },
                "email": {
                  "type": "string",
                  "example": "joao@email.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Cliente atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        }
      }
    },
...
    "/pedidos/": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Lista os pedidos do cliente autenticado",
        "description": "Retorna apenas os pedidos do cliente autenticado.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Cria um novo pedido",
        "description": "Permite a criação de um novo pedido no sistema.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido criado com sucesso."
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Obtém um pedido pelo ID",
        "description": "Retorna os detalhes de um pedido específico com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido retornado com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pedido não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/administrador/pedidos": {
      "get": {
        "tags": [
          "Pedidos Admin"
        ],
        "summary": "Consulta todos os pedidos",
        "description": "Consulta todos os pedidos registrados.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Pedidos consultados com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/administrador/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos Admin"
        ],
        "summary": "Consulta um pedido",
        "description": "Consulta os dados de um pedido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Pedido consultado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedidos Admin"
        ],
        "summary": "Exclui um pedido",
        "description": "Exclui um pedido existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido excluído com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/itens-pedido/": {
      "get": {
        "tags": [
          "Itens de Pedido"
        ],
        "summary": "Lista todos os itens de pedido",
        "description": "Retorna uma lista com todos os itens de pedido cadastrados.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Itens de pedido listados com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/itens-pedido/{id}": {
      "get": {
        "tags": [
          "Itens de Pedido"
        ],
        "summary": "Retorna um item de pedido específico",
        "description": "Retorna os detalhes de um item de pedido com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item de pedido retornado com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item de pedido não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Itens de Pedido"
        ],
        "summary": "Exclui um item de pedido",
        "description": "Exclui um item de pedido com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item de pedido excluído com sucesso."
          },
          "400": {
            "description": "Não foi possível excluir o item de pedido."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}